LayoutTests/media/video-playback-system-interruption-expected.txt


RUN(internals.setMediaElementRestrictions(video, 'NoRestrictions'))
RUN(internals.beginMediaSessionInterruption("EnteringBackground"))
RUN(internals.beginMediaSessionInterruption("System"))
RUN(video.volume = 0.1)
RUN(video.src = findMediaFile("video", "content/audio-tracks"))
EVENT(canplaythrough)
EXPECTED (video.paused == 'true') OK
EXPECTED (internals.mediaSessionState(video) == 'Interrupted') OK
RUN(internals.endMediaSessionInterruption("MayResumePlaying"))
EVENT(playing)
EXPECTED (internals.mediaSessionState(video) != 'Interrupted') OK
END OF TEST
